+++ libs/web/ui/src/index.js
@@ -1,3 +1,6 @@
+export * from './Admin/RatingList'
+export * from './Admin/SpellcheckedList'
+export * from './Admin/UserList'
 export * from './Wabap/WabapProvider'
 export * from './Wabap/SearchBox'
 export * from './Components/FeatherIcon'